% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{vacc_calc_response_summed}
\alias{vacc_calc_response_summed}
\alias{vacc_calc_response_profile}
\alias{vacc_calc_fds_prop}
\title{Calculate the summed and profile response}
\usage{
vacc_calc_response_summed(.data, cn_resp, grp = NULL)

vacc_calc_response_profile(.data, cn_resp, grp = NULL)

vacc_calc_fds_prop(.data, cn_resp, grp = NULL)
}
\arguments{
\item{.data}{A data frame.}

\item{cn_resp}{character.
Column name in \code{.data} that contains the response.}

\item{grp}{character vector.
Columns in \code{.data} containing the grouping variable.
If \code{NULL} (the default), then these are automatically determined.}
}
\description{
The functions \code{calc_response_summed} and \code{calc_response_profile}
calculate the summed and profile response, respectively,
as defined in Rodo (2019).
}
