% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calcLU}
\alias{calcLU}
\title{Calculate number of unique vector elements.}
\usage{
calcLU(x)
}
\arguments{
\item{x}{vector.}
}
\description{
Convenience function around \code{length()} and \code{unique()}
that calculates the number of unique elements in a vector.
}
