% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calcSub}
\alias{calcSub}
\title{Calculate baseline-subtracted response.}
\usage{
calcSub(dataTbl)
}
\arguments{
\item{dataTbl}{dataframe. A dataframe with least the columns med, timePoint and resp, containing
the longitudinal responses for a single individual for a single response type.}
}
\value{
A dataframe with either that individual's day zero response subtracted from all responses,
or the median response at day for that individual's group subtracted from all responses.
}
\description{
Calculate baseline-subtracted response.
}
