% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{changeCytOrder}
\alias{changeCytOrder}
\title{Change order of cytokines in cytokine combination labels.}
\usage{
changeCytOrder(
  dataTibble,
  firstCytColIndex,
  firstCytNameLoc,
  signLocVec,
  orderLocVec,
  cdPresent = TRUE
)
}
\arguments{
\item{dataTibble}{dataframe. A dataframe with the cytokine
combination frequency response variables the last (right-most)
variables.}

\item{firstCytColIndex}{integer. The index of the first column
containing cytokine combination response data.}

\item{firstCytNameLoc}{integer. The first position in the string
of the cytokine combination label that specifies a cytokine.
This helps avoid prefixes like 'CD4' or 'CD8'.}

\item{signLocVec}{integer vector. The positions of the positivity
indicators for each cytokine.}

\item{orderLocVec}{integer vector. The i-th entry denotes the desired
order of the cytokine combination that originally appeared
i-th.}

\item{cdPresent}{logical. If TRUE, then the cytokine combination
originally is taken to be prefixed with CD4 or CD8, and
the cluster of differentiation is then stored as a
character vector. Defaults to TRUE.}
}
\value{
A dataframe with the order of the cytokines
in the cytokine combinations changed.
}
\description{
\code{changeCytOrder} changes the order of cytokine combinations
in a dataframe.
}
