% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{compToFullName}
\alias{compToFullName}
\title{Compare expected to actual column entries for BCG source data.}
\usage{
compToFullName(dataTibble, compColumnName, mainColumnName)
}
\arguments{
\item{dataTibble}{dataframe. A dataframe containing columns with
names as per \code{mainColumnName} and \code{compColumnName}.}

\item{mainColumnName, compColumnName}{character. Strings denoting the
actual (\code{mainColumnName}) and expected (\code{compColumnName})
entries in a row of \code{dataTibble}.}
}
\value{
A list of the differing entries in \code{mainColumnName} and
\code{compColumnName}.
}
\description{
\code{compToFullName} provides a list of rows for which the
expected data, as per \code{mainColumnName}, is not found in
\code{compColumnName}.
}
