% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{convProtNameToInteger}
\alias{convProtNameToInteger}
\title{Convert protocol name to assigned integer.}
\usage{
convProtNameToInteger(protVec)
}
\arguments{
\item{protVec}{character vector. Vector containing protocol names.}
}
\value{
character vector.
}
\description{
\code{convProtNameToInteger} converts the names of the
trial protocols
each individual was given, as per the source file, to assigned
integers (stored as characters).
}
\details{
The conversion table is as follows.
\tabular{cc}{
Original name \tab Converted name \cr
402.003 \tab 1 \cr
402.032 \tab 2 \cr
GSK.010 \tab 3 \cr
GSK.012 \tab 4 \cr
MVA.008 \tab 5 \cr
MVA.011 \tab 6 \cr
THYB-04 \tab 7 \cr
#' TBRU \tab 8 \cr
}
}
