% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{convTimePoint}
\alias{convTimePoint}
\title{Converts the original timepoints in TBRU to days.}
\usage{
convTimePoint(studyDayVector)
}
\arguments{
\item{studyDayVector}{character vector. The character vector
based on the column in the original full data table that contains
the timepoint information.}
}
\description{
\code{convTimePoint} converts the codes for timepoints in the
original TBRU dataset to days.
}
\details{
The conversions were done as follows.

\tabular{cc}{
Original value \tab Converted value \cr
V08 \tab 0 \cr
V13 \tab 21 \cr
V15 \tab 35 \cr
V28 \tab 365 \cr
}
}
