% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{convVaccNameToInteger}
\alias{convVaccNameToInteger}
\title{Convert vaccine name to assigned integer.}
\usage{
convVaccNameToInteger(vaccVec)
}
\arguments{
\item{vaccVec}{character vector. Vector containing vaccine names.}
}
\value{
character vector.
}
\description{
\code{convVaccNameToInteger} converts the names of the vaccines
each individual was given, as per the source file, to assigned
integers (stored as characters).
}
\details{
The conversion table is as follows.
\tabular{cc}{
Original name \tab Converted name \cr
402 \tab 1 \cr
H56 \tab 2 \cr
GSK.M72 \tab 3 \cr
MVA85A \tab 4 \cr
H1 \tab 5 \cr
BCG \tab 6 \cr
}
}
