% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{countFuncClosure}
\alias{countFuncClosure}
\title{A functional operator that returns a counting function.}
\usage{
countFuncClosure()
}
\value{
A function as described in the description.
}
\description{
Returns a function that returns an integer
one larger than the last integer it returned, starting
from one. In other words, the first time the
outputted function is run, it returns one, the second
time two, the third time three, and so on.
}
