% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{vacc_extract_baseline}
\alias{vacc_extract_baseline}
\alias{vacc_extract_peak}
\alias{vacc_extract_memory}
\title{Extract data from particular timepoints}
\usage{
vacc_extract_baseline(.data, cn_time = "timepoint")

vacc_extract_peak(.data, cn_time = "timepoint", peak_time = NULL)

vacc_extract_memory(.data, cn_time = "timepoint", mem_time = NULL)
}
\arguments{
\item{.data}{A data frame.}

\item{cn_time}{character.
Column name in \code{.data} that contains the timepoint.}

\item{peak_time}{named numeric vector.
Named vector specifying the peak timepoint for each vaccine.
Names are the protocol names, and values are the peak timepoints.
Must be supplied}

\item{mem_time}{named numeric vector.
Named vector specifying the memory timepoint for each vaccine.
Names are the protocol names, and values are the memory timepoints.
If not supplied, then the defaults are used.}
}
\description{
These functions extract data from particular timepoints.
The functions \code{vacc_extract_baseline}, \code{vacc_extract_peak}, and
\code{vacc_extract_memory} extract the baseline, peak, and memory
response, respectively.
The timepoints are specified in the \code{peak_time} and \code{mem_time}
arguments for the \code{vacc_extract_peak} and \code{vacc_extract_memory}
functions, respectively.
}
