% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{nameStim}
\alias{nameStim}
\title{Give stims for each vaccine}
\usage{
nameStim(x)
}
\description{
Given a vector of vaccine number references, it
returns the stim for that vaccine. Designed
for the lOrigMainTbl.
}
