% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{removeNonANEntries}
\alias{removeNonANEntries}
\title{Remove the non-alphanumeric entries in specified columns.}
\usage{
removeNonANEntries(dataTibble, columnNameVec)
}
\arguments{
\item{dataTibble}{dataframe. A dataframe containing the
columns in columnNameVec.}

\item{columnNameVec}{character vector. A character vector
of the columns in \code{dataTibble} for which the
non-alphanumeric entries must be removed.}
}
\description{
\code{removeNonANEntries} removes the non-alphanumeric entries in
columns specified by \code{columnNameVec} in \code{dataTibble}.
}
