% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{replaceCytPosIndicator}
\alias{replaceCytPosIndicator}
\title{Replace cytokine combination positivity indicator.}
\usage{
replaceCytPosIndicator(
  dataTibble,
  firstMultCytIndex,
  posPattern,
  negPattern,
  posRep,
  negRep,
  alphaNumericOrigSign = FALSE,
  signLocVec = NULL
)
}
\arguments{
\item{dataTibble}{dataframe. A dataframe with the last columns
only being the names of cytokine combinations.}

\item{firstMultCytIndex}{integer. The column index of the first
column for frequencies for a cytokine combination.}

\item{posPattern, negPattern}{character. The original indicators for positivity
and negativity.}

\item{posRep, negRep}{character. The replacement indicators for
positivity and negativity.}

\item{alphaNumericOrigSign}{logical. If TRUE, then posPattern and
negPattern are alphanumeric.}

\item{signLocVec}{integer vector. Specifies the character positions of
posPattern and negPattern.}
}
\value{
A dataframe with cytokine combination labels formatted to
having the chosen cytokine positivity indicators.
}
\description{
\code{replaceCytPosIndicator} replaces whatever the original
indicators were for cytokine positivity and negativity in
the cytokine combination shorthand labels with selected
replacements (such as 'p' and 'n').
}
