% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_chunk_reorder.R
\name{str_locate_multi}
\alias{str_locate_multi}
\title{Locate all occurrences of all strings in a character vector.}
\usage{
str_locate_multi(string, pattern)
}
\arguments{
\item{string}{character. String to search through.}

\item{pattern}{character vector. Vector of strings to search for in \code{string}.}
}
\value{
Numeric vector of found positions for any character in \code{pattern},
sorted in ascending order.
}
\description{
\code{str_locate_multi} provides the a sorted vector
of the locations of character elements (or strings) in a character
string.
}
\examples{
str_locate_multi("abcdeabcdexyza", c("ab", "x"))
}
