% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_chunk_reorder.R
\name{str_split_chunk}
\alias{str_split_chunk}
\title{Split a string into chunks.}
\usage{
str_split_chunk(string, pattern)
}
\value{
A list of strings.
}
\description{
\code{str_split_chunk} works like str_split, except that it:
\itemize{
\item allows detection of multiple characters, and
\item includes the characters from \code{pattern} in the
chunks.
}
}
\examples{
str_split_chunk("GpTpIL2nIL17p", c("p", "n"))
}
