% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_chunk_reorder.R
\name{str_split_chunk_pos}
\alias{str_split_chunk_pos}
\title{Find the border positions of the characters to split for.}
\usage{
str_split_chunk_pos(string, pattern)
}
\arguments{
\item{string}{character. String to search through.}

\item{pattern}{character vector. Vector of strings to search for in \code{string}.}
}
\value{
A list with each element an ordered integer
vector containing the start and end position of each
chunk.
}
\description{
\code{str_split_chunk_pos} provides a list, where
each element is a length-two integer vector of the
locations at which \code{string} must be split at
to create sub-strings (or chunks), where the right-hand
endpoints of thhe chunks are at an element of
\code{pattern}.
}
\examples{
str_split_chunk_pos("abcdeabcdexyza", c("a", "x"))
}
