% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_to_upper_sel.R
\name{str_to_upper_sel}
\alias{str_to_upper_sel}
\title{Capitalise all letters in a string except some using rename.}
\usage{
str_to_upper_sel(x, vec1)
}
\arguments{
\item{x}{character vector. Character vector of column names.
Provided by rename.}

\item{vec1}{character vector. Character vector of characters
to not capitalise.}
}
\description{
\code{str_to_upper_sel_rename}
}
