% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_to_upper_sel.R
\name{str_to_upper_sel_elem}
\alias{str_to_upper_sel_elem}
\title{Capitalise a character if it is not in a vec1.}
\usage{
str_to_upper_sel_elem(x, vec1)
}
\arguments{
\item{x}{character. Character to possibly capitalise.}

\item{vec1}{character vector. Vector of characters to not
capitalise.}
}
\value{
A character that is capitalised if it not in vec1,
and is not capitalised if it is in vec1.
}
\description{
Capitalises an individual character if is
not in an exclusion vector. Only used in \code{str_to_upper_sel},
and not exported.
}
