% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_to_upper_sel.R
\name{str_to_upper_sel_string}
\alias{str_to_upper_sel_string}
\title{Capitalise all letters in a string except some.}
\usage{
str_to_upper_sel_string(x, vec1)
}
\arguments{
\item{x}{character. String to}

\item{vec1}{character vector. Characters against which each individual
part of x is compared}
}
\value{
A character string.
}
\description{
\code{str_to_upper_sel} capitalises all letters in \code{x},
except for those specified in \code{vec1}.
}
\examples{
str_to_upper_sel("abc", NULL)
str_to_upper_sel("abc", c("b", "c"))
}
