% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{strConv}
\alias{strConv}
\alias{editStr}
\alias{editNonANStr}
\alias{repPosNegInd}
\title{String convenience functions}
\usage{
editStr(str1, startLoc, endLoc)

editNonANStr(str1, startLoc, endLoc)

repPosNegInd(str1)
}
\arguments{
\item{str1}{character.}

\item{startLoc, endLoc}{integer.}
}
\description{
Wrapper functions around str_ functions.
}
