% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{vacc_calc_vaccine_induced_response}
\alias{vacc_calc_vaccine_induced_response}
\title{Calculate the vaccine-induced response}
\usage{
vacc_calc_vaccine_induced_response(.data, cn_resp, grp = NULL)
}
\description{
To calculate the vaccine-induced response, we subtract the baseline response.
If the baseline response is missing for a particular individual,
we subtract the
median baseline response for that individual's vaccine and
infection status group.
}
