% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{vacc_set_neg_to_zero}
\alias{vacc_set_neg_to_zero}
\title{Set negative responses to zero}
\usage{
vacc_set_neg_to_zero(.data, cn_resp)
}
\arguments{
\item{.data}{A data frame.}
}
\description{
Set negative responses to zero
}
